requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.7.0.72
dlib-bin==19.24.6
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0